Add `filter_accessible_notes_by` method to `Note`
This is in turn an easy way to only query for notes that a user can
access, either because they are the owner, or because it was shared with
them. This provides a better spearation of concerns and avoids querying
unnecessary data.

However this method doesn't care if the user is an editor or only a
reader of the given note. This has to be checked manually.

Query implemented as a method, extending the `QuerySet` of the `Note`
model. Replaced `Note`'s `objects` `Manager`.

Added shortcut function for getting one note, accessible by user or
raising  404 if that doesn't exist.