Added user search API endpoint
Designed it in such a way to not leak the entire userbase list.

If there are more than 5 results, it shows nothing, unless one of them
is an exact match, in which case it only returns the exact match.

If search query is less than 3 chars, it only returns exact match, if
any.