Added SharedItemListView
On GET, it returns a list of `SharedItem`s of the current note.

On POST, it adds a new `SharedItem` to the note, or updates existing one
if any.

On PUT, it sets the `SharedItem`s to the given list

On PATCH, it extends the `SharedItem`s with the given list

On DELETE, it clears the `SharedItem`s