Installed all the Svelte Material UI components
I don't know why somebody thought it's a good idea to install each
component as separate package, but I don't like it.

Maybe it IS a good idea and if that's the case, I will revert this commit.

I also had to add TypeScript support, because the new version of svelte
material ui needed it.