Made only verified notes show in notes index page
This is done to prevent spam mostly.

Made publish happen on POST, instead of GET. This mitigates the
potantial CSRF attack where a user unknowingly gets a note of theirs
published because they clicked a link or loaded an image with the source
set to the `/publish/` route.

Also made it so notes get verified automatically if published by staff
user. We may or may not regret this later.